requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
marshmallow
flask-marshmallow
python-dotenv
flask-migrate
SQLAlchemy
flask-smorest
Werkzeug
flask-jwt-extended